♻️ refactor: toml parsing of confg
